Extract inline plugin specs into lua/custom/plugins/
Plugin configuration was split between two inconsistent locations: most
specs inline in init.lua, a few in separate kickstart modules, with no
rule for which went where. init.lua had grown to ~900 lines.

Move each inline spec to its own file under lua/custom/plugins/, picked up
automatically by the custom.plugins import. init.lua now holds only
options, keymaps, autocommands and the lazy bootstrap (~300 lines).

Specs moved verbatim (telescope, lsp, blink-cmp, conform, treesitter,
mini, colorscheme, todo-comments, which-key, guess-indent); no behaviour
changes for those.

gitsigns is the exception: its inline `signs` block is dropped here, so
only the kickstart module's keymaps remain until the next commit restores
the signs config in a single consolidated spec.
