+++ lua/custom/plugins/guess-indent.lua
@@ -1,0 +1,6 @@
+-- Detect tabstop and shiftwidth automatically
+-- https://github.com/NMAC427/guess-indent.nvim
+
+---@module 'lazy'
+---@type LazySpec
+return { 'NMAC427/guess-indent.nvim', opts = {} }
