+++ lua/custom/plugins/todo-comments.lua
@@ -1,0 +1,14 @@
+-- Highlight todo, notes, etc in comments
+-- https://github.com/folke/todo-comments.nvim
+
+---@module 'lazy'
+---@type LazySpec
+return {
+  'folke/todo-comments.nvim',
+  event = 'VimEnter',
+  dependencies = { 'nvim-lua/plenary.nvim' },
+  ---@module 'todo-comments'
+  ---@type TodoOptions
+  ---@diagnostic disable-next-line: missing-fields
+  opts = { signs = false },
+}
